[cxx-interop] Disable failing tests on Amazon Linux 2
